feat: add parameter to keep the outdir before bundling starts

+++ src/lib/bundle.js
@@ -5,14 +5,21 @@ import path from 'path';
 import shell from 'shelljs';
 import winston from 'winston';
 
-export default function compileBundle({ dir, workingDir = tmp.dirSync().name, ci } = {}) {
+export default function compileBundle({
+  dir, workingDir = tmp.dirSync().name, ci, keep,
+} = {}) {
   const customMeteorProjectDirShellEx = `cd ${dir} &&`;
 
   winston.info('Compiling application bundle');
 
   // Generate Meteor build
   winston.debug(`generate meteor build at ${workingDir}`);
-  shell.exec(`rm -rf ${workingDir}`);
+  if (!keep) {
+    winston.debug(`removing ${workingDir}`);
+    shell.exec(`rm -rf ${workingDir}`);
+  } else {
+    winston.debug(`keeping ${workingDir}, if it exists`);
+  }
   shell.exec(`${dir ? customMeteorProjectDirShellEx : ''} meteor build ${workingDir} ${ci ? '--allow-superuser' : ''} --directory --server-only --architecture os.linux.x86_64`);
 
   // Cleanup broken symlinks
